requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
attrs==22.1.0
    # via
    #   cattrs
    #   jsii
autopep8==1.7.0
    # via python-cdk-starter-kit (pyproject.toml)
aws-cdk-aws-lambda-python-alpha==2.39.1a0
    # via python-cdk-starter-kit (pyproject.toml)
aws-cdk-lib==2.39.1
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   python-cdk-starter-kit (pyproject.toml)
cattrs==22.1.0
    # via jsii
constructs==10.1.92
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   python-cdk-starter-kit (pyproject.toml)
datadog-cdk-constructs-v2==0.3.5

exceptiongroup==1.0.0rc9
    # via cattrs
jsii==1.66.0
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
pycodestyle==2.9.1
    # via autopep8
python-dateutil==2.8.2
    # via jsii
python-dotenv==0.20.0
    # via python-cdk-starter-kit (pyproject.toml)
six==1.16.0
    # via python-dateutil
toml==0.10.2
    # via autopep8
typeguard==2.13.3
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.3.0
    # via jsii
